refactor: remove word import handlers, singleton pattern, and bump version
- Bumped package version from 1.4.8 to 1.4.10
- Removed singleton pattern from CustomModelDropdown by making constructor public and commenting out getInstance method
- Cleaned up formatting and trailing spaces in EverEtchApp.ts
- Refactored ModalManager.ts by removing import-related methods, handlers, and unused imports to simplify the component and reduce complexity
