+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "everetch",
-  "version": "1.4.8",
+  "version": "1.4.10",
   "type": "module",
   "description": "A slim word memo application",
   "main": "lib/main.js",
